requirements.txt
openai>=1.30.0
requests>=2.31.0
fastapi>=0.110.0
uvicorn>=0.29.0
neo4j>=5.20.0
hnswlib>=0.8.0
sentence-transformers>=2.6.0
numpy>=1.26.0
pydantic>=2.6.0
mcp>=1.0.0
httpx>=0.27.0
python-dotenv>=1.0.1
dashscope>=1.20.0
websockets>=12.0